Random stuff

+++ MixedMediaInventoryTracker/Models/SoldMediaItemDetailsModel.cs
@@ -9,7 +9,7 @@ namespace MixedMediaInventoryTracker.Models
     {
         public int Id { get; set; }
         public string Buyer { get; set; }
-        public int? Amount { get; set; }
+        public float? Amount { get; set; }
         public DateTime SoldDate { get; set; }
         public string Notes { get; set; }
         public string Title { get; set; }
